Impoved logging and command line options.
`VERBOSITY` (previously just `VERBOSE`) is now handled as 6 levels for
the purpose of logging:
- `VERBOSITY_ERROR` Errors that cause program to quit immediately
- `VERBOSITY_WARNING` Errors that do not cause program to quit
  immediately
- `VERBOSITY_NOTICE` A notable action taken by the script
- `VERBOSITY_INFO` Information about script flow
- `VERBOSITY_CONFIG` Information about configurations that have been
  applied
- `VERBOSITY_DEBUG` Full debugging information

All logging (after the logging is initialised) goes through the log
function which automatically compares the set `VERBOSITY` level with
the log messages verbosity.

The default verbosity is `VERBOSITY_WARNING`. This can be set back to
`VERBOSITY_ERROR` using the `-q` or `--quiet` flags that have been
added or increased using the `-v` or `--verbose` flags. The `--debug`
flag can be used to immediately set the `VERBOSITY` to
`VERBOSITY_DEBUG` and enable other debugging facilities. Resolves #15.

Finally a `-c`/`--config` flag has been added to set the config
directory. Resolves #14.