Use exclude default to exclude config files.
Rather than using a minimum depth of 2 to exclude files in the root of
the config directory, the configuration files related to
shared-config-init are listed in the `exclude-default` file. Resolves #19.

exclude-default will now fall back to using the one stored in
`SCRIPT_DIR` instead of requiring it to be in the `CONFIG_DIR` and now
makes the existence in the `CONFIG_DIR` work as an override.